refactor: move SendToNetwork functions to utils file
